requirements.txt
streamlit==1.39.0
opencv-python==4.10.0.84
Pillow==10.4.0
ultralytics==8.2.95
numpy==1.26.4
pandas==2.2.2
torch>=2.0.0
torchvision>=0.15.0
matplotlib==3.9.2
tqdm==4.66.5
pyyaml==6.0.2
datetime
